refactor: 트랜잭션 로직 분리를 위한 UserService 인터페이스 도입 및 구현체 분리
