+++ app/src/main/java/com/techcourse/service/UserService.java
@@ -1,77 +1,12 @@
 package com.techcourse.service;
 
-import com.interface21.dao.DataAccessException;
-import com.interface21.jdbc.datasource.DataSourceUtils;
-import com.interface21.transaction.support.TransactionSynchronizationManager;
-import com.techcourse.dao.UserDao;
-import com.techcourse.dao.UserHistoryDao;
 import com.techcourse.domain.User;
-import com.techcourse.domain.UserHistory;
-import java.sql.Connection;
-import java.sql.SQLException;
-import javax.sql.DataSource;
-import org.slf4j.Logger;
-import org.slf4j.LoggerFactory;
 
-public class UserService {
+public interface UserService {
 
-    private static final Logger log = LoggerFactory.getLogger(UserService.class);
+    User findById(final long id);
 
-    private final UserDao userDao;
-    private final UserHistoryDao userHistoryDao;
-    private final DataSource dataSource;
+    void save(final User user);
 
-    public UserService(final UserDao userDao, final UserHistoryDao userHistoryDao, final DataSource dataSource) {
-        this.userDao = userDao;
-        this.userHistoryDao = userHistoryDao;
-        this.dataSource = dataSource;
-    }
-
-    public User findById(final long id) {
-        return userDao.findById(id);
-    }
-
-    public void insert(final User user) {
-        userDao.insert(user);
-    }
-
-    // TODO 4단계에서 트랜잭션 로직 분리하기
-    public void changePassword(final long id, final String newPassword, final String createBy) {
-        Connection conn = null;
-        try {
-            conn = DataSourceUtils.getConnection(dataSource);
-            conn.setAutoCommit(false);
-            try {
-                User user = userDao.findById(id);
-                user.changePassword(newPassword);
-                userDao.update(user);
-                userHistoryDao.log(new UserHistory(user, createBy));
-                conn.commit();
-            } catch (Exception e) {
-                rollback(conn, e);
-                throw e;
-            }
-        } catch (SQLException e) {
-            log.error(e.getMessage(), e);
-            throw new DataAccessException(e);
-        } finally {
-            TransactionSynchronizationManager.unbindResource(dataSource);
-            try {
-                if (conn != null) {
-                    conn.close();
-                }
-            } catch (SQLException e) {
-                log.error("Failed to close connection", e);
-            }
-        }
-    }
-
-    private void rollback(Connection conn, Exception originalException) {
-        try {
-            conn.rollback();
-        } catch (SQLException rollbackEx) {
-            log.error("Rollback failed", rollbackEx);
-            originalException.addSuppressed(rollbackEx);
-        }
-    }
+    void changePassword(final long id, final String newPassword, final String createdBy);
 }
